// apps/web/src/routes/_dashboard/$user/repos/$repo/-components/__generated__/Branches_refs.graphql.ts
export type Branches_refs$data = {
  readonly id: string;
  readonly refs: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
        readonly target: {
          readonly commits: {
            readonly " $fragmentSpreads": FragmentRefs<"Commits_history">;
          } | null | undefined;
        } | null | undefined;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  } | null | undefined;
  readonly " $fragmentType": "Branches_refs";
};
export type Branches_refs$key = {
  readonly " $data"?: Branches_refs$data;
  readonly " $fragmentSpreads": FragmentRefs<"Branches_refs">;
};